LayoutTests/platform/chromium/accessibility/footer-roledescription-expected.txt

footer
content info
This tests that a footer element has a different role description from a contentinfo

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS footer.role is contentInfo.role
PASS footer.subrole is contentInfo.subrole
PASS footer.roleDescription is 'AXRoleDescription: footer'
PASS contentInfo.roleDescription is 'AXRoleDescription: content information'
PASS successfullyParsed is true

TEST COMPLETE
